+++ src/main/java/com/example/springpj/admin/service/AdminService.java
@@ -51,13 +51,13 @@ public class AdminService {
     }
 
     //회원 상세 조회 중간 처리
-    public User getUserContent(int csNum) {
+    public User getUserContent(Long csNum) {
         log.info("회원 상세 조회 중간 처리! " +csNum);
         return adminMapper.getUserContent(csNum);
     }
 
     //AS 상세 조회 중간 처리
-    public Request getRequestContent(int asNum) {
+    public Request getRequestContent(Long asNum) {
         log.info("회원 상세 조회 중간 처리! " +asNum);
         return adminMapper.getRequestContent(asNum);
     }
